Fix "development version tests" by using terra pre-release as well
As of qiskit-nature commit b7cbbf63569acf8f9a7, it now depends on
Terra 0.22 or higher, which has not yet been released.  So if we
want to continue testing on the development version of qiskit-nature,
we are forced to test on the development version of qiskit-terra
as well.  The down side of this is that it may eat our CI
minutes a bit more, since qiskit-terra takes longer to
compile and install from git.