Fix "development version tests" by using terra pre-release as well (#7)
* Fix "development version tests" by using terra pre-release as well

As of qiskit-nature commit b7cbbf63569acf8f9a7, it now depends on
Terra 0.22 or higher, which has not yet been released.  So if we
want to continue testing on the development version of qiskit-nature,
we are forced to test on the development version of qiskit-terra
as well.  The down side of this is that it may eat our CI
minutes a bit more, since qiskit-terra takes longer to
compile and install from git.

* Fix Estimator for use with Qiskit Terra 0.22

The `BaseEstimator` sets `__hash__ = None` as of Qiskit Terra
0.22, so we cannot pass it to the `lru_cache` (which happens
via `self` when the `_build_statevector` method is called).

* Satisfy mypy