Set `platform: linux/amd64` in `docker-compose.yml` (#71)
By default, Docker running on a Mac M1 chip will create a container that is native to M1.  However, for the purposes of the toolkit, we need to install cplex, which only is built for Intel.  So we set the platform explicitly in the `docker-compose.yml` file, which tells the M1 chip to use its `x86_64` emulation instead.