Require 100% test coverage in certain modules
Right now, I have enabled the check for 100% coverage in a few
modules that already have it.  Going forward, we plan to also use
this check for new modules.  When we introduce new modules that
should have 100% coverage, we will have to add their paths to
`tox.ini` using globs (e.g.,
`circuit_knitting_toolbox/circuit_cutting/utils/**/*`), following
https://coverage.readthedocs.io/en/7.2.3/source.html#file-patterns.

The old coverage check, of all modules, follows in the coverage
workflow, so that we can report the correct number to coveralls.

Closes #107.