Remove any attempt at using the samplers in parallel (#207)
This is an alternative to #206, which was still causing sporadic
failures for me locally.  I just tested this current change
a number of times, and there were zero hangs.

If this indeed works consistently, then I suspect we can
come up with a minimal test case that demonstrates the
hang we were experiencing with Aer's Sampler and `PrimitiveJob`
in Terra.