Bump qiskit-ibm-runtime version and use `mode=batch` (backport #623)
This is necessary to prevent a `DeprecationWarning` under
qiskit-ibm-runtime 0.24, which is currently visible in the
[first](https://qiskit-extensions.github.io/circuit-knitting-toolbox/circuit_cutting/tutorials/01_gate_cutting_to_reduce_circuit_width.html)
and third tutorials and is due to the merging of
https://github.com/Qiskit/qiskit-ibm-runtime/pull/1584.

This means that the notebooks now depend on
qiskit-ibm-runtime 0.24, which depends on qiskit>=1.1.

The "minimum version tests" aren't equipped to deal with situations
where the notebooks have a different minimum version than the
library.  For this reason I've disabled the "minimum version tests"
on the current stable branch going forward for the notebooks only.
I don't want to bump the minimum version of any library on the
stable branch, but I _do_ want to update our code examples
to the latest version of qiskit-ibm-runtime.