Pin `ruff==0.5.1`
We pin most of our linters, but we have gotten away so far without
pinning ruff.  However, today CI is broken because the ruff 0.5.2
(patchlevel) release seems to have added a rule.  (Or maybe it
fixed a rule that was broken before.  Either way, it broke us.)

From here forward, let's pin ruff, too, and dependabot can propose
updating the version from time to time, just as it does with the
other linters.