Convert inline math into displayed equation to avoid overflowing in the future (#759)
This PR converts an inline math expression into a displayed equation. Soon, we will be changing how the inline expressions are showed on the website and we need to transform long expressions to have a better visualization of them and avoid any possible overflow.

Even though, the current expression looks good on the website right now, it will soon break if we don't apply this change. 

This is an example of the final render:

<img width="825" alt="Screenshot 2025-06-25 at 10 30 25" src="https://github.com/user-attachments/assets/370f02d8-5b64-4f8d-b2fb-cf4bdc5b8208" />